Create equal.js
demonstrates the use of the equal() function in nodejs assert module. It uses the == operator to check whether two values passed into the function are equal or not. See that it returns true in the case when a integer 50 is said to be equal with string 50. This is because the equality in this case is not strict